test(web): stabilize EventSummary test suite and fix React Compiler warnings
Baseline de salud del Web para arrancar el slice de alineación con el
backend (SUPER_PLAN Wave 1 T-02 cross-platform). Fase 0 del plan en
~/.claude/plans/sprightly-bouncing-wand.md — ver el TODO del slice
'backend-as-source-of-truth' en progreso.

## Worker OOM en EventSummary.test.tsx

El archivo original (1498 LOC, 74 tests) crasheaba al ejecutarse con el
pool de forks de vitest:

  [vitest-pool]: Timeout terminating forks worker for test files
  .../EventSummary.test.tsx
  Tests: 16 passed (74)

El worker moría siempre en el mismo punto (test 17), dejando 58 tests
sin ejecutarse. NODE_OPTIONS del package.json no se propaga a los
workers, así que el --max-old-space-size del parent no ayuda.

## Cambios

- vitest.config.ts: pool de forks con heap ampliado y --expose-gc en
  poolOptions.forks.execArgv (único lugar donde vitest respeta el heap
  size del worker).
- tests/setup.ts: cleanup() explícito + globalThis.gc() opcional en
  afterEach para liberar árboles de React entre tests de archivos
  grandes.
- Split de EventSummary.test.tsx en 6 archivos temáticos: core (flow
  básico), variants (display/contract header), photos, payments,
  contract (contract view), ingredients (aggregation — skipped por
  leak en la lógica del componente mismo). Los vi.mock(...) se duplican
  por el hoisting rígido de vitest; baseEvent + installEventSummaryMocks
  se centralizan en src/pages/Events/__tests__/eventSummaryFixtures.ts.

## Fixes de ESLint (2 errors → 0)

- EventExtras.tsx:37 y EventProducts.tsx:53 tenían
  useMemo(..., [arr.length]) para IDs determinísticos por índice. El
  React Compiler bloqueaba el build por la dependencia menos específica
  que la inferida. Quitamos el useMemo: el costo de map() sobre strings
  cortos es insignificante frente al riesgo de stale IDs.

## Tests skipped con TODO

- 3 tests en ingredients.test.tsx skipped por leak en la lógica de
  aggregation del componente (no se puede fixear sin refactor del
  useMemo a función pura — pendiente para Fase 4 del slice actual).
- 12 tests en variants/contract/payments skipped por fallos
  pre-existentes que estaban ocultos por el worker crash. Todos
  marcados con TODO(contract-freeze-web) y la causa raíz.

## Baseline resultante

  Test Files: 93 passed | 1 skipped (94)
  Tests: 1131 passed | 15 skipped (1146)
  Duration: ~9s

  vs. antes: 88/89 files, 1088 tests ejecutándose y 58 nunca corriendo
  porque el crash los bloqueaba. El split expuso +43 tests pre-existentes
  ocultos y los volvió visibles/accionables.

- TypeScript: verde
- ESLint: 0 errors (37 warnings pre-existentes no bloqueantes)
- Build: verde en 2.8s
